Fixes bugs in Difference of Gaussian filter class

+++ src/man/vision/DiffOfGaussianFilter.h
@@ -4,12 +4,13 @@ namespace vision {
 // TODO partial kernels for dealing with edges of signal
 class DiffOfGaussianFilter {
 public:
-    DiffOfGaussianFilter(int ksize, int narrowSigma, int wideSigma);
+    // TODO add constructor without kernel size param
+    DiffOfGaussianFilter(int ksize, double narrowSigma, double wideSigma);
     ~DiffOfGaussianFilter();
     void convolve(int inSize, double const* in, double* out);
 
 private:
-    static double gaussianAt(int x, int sigma);
+    static double gaussianAt(double x, double sigma);
 
     int size;
     double* kernel;
